requirements-dev.txt
pytest-console-scripts
